+++ ft_low_hex_print.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_low_hex_print.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttachi <ttachi@student.42tokyo.jp>         +#+  +:+       +#+        */
+/*   By: ttachi <ttachi@student.42tokyo.ja>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 17:40:45 by ttachi            #+#    #+#             */
-/*   Updated: 2022/11/03 20:37:44 by ttachi           ###   ########.fr       */
+/*   Updated: 2022/11/04 19:43:37 by ttachi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 int	ft_low_hex_print(const char **argv, va_list ap)
 {
-	unsigned int	x;
-	int				digit;
+	unsigned long long	x;
+	int					digit;
 
 	digit = 1;
-	x = va_arg(ap, unsigned int);
+	x = va_arg(ap, unsigned long long);
 	digit = ft_hex_print(x, LOWERCASE_HEX, digit);
 	argv++;
 	return (digit);
